pyproject.toml
[tool.poetry]
name = "basetopia-be"
version = "0.1.0"
description = "Baseball Statistics and Analysis API"
authors = ["Michael Guyon <mikeguyon98@gmail.com>"]
readme = "README.md"
packages = [{include = "basetopia_be"}] 

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
firebase-admin = "^6.6.0"
python-dotenv = "^1.0.1"
google-cloud-firestore = "^2.20.0"
requests = "^2.32.3"
pytest = "^8.3.4"
beautifulsoup4 = "^4.12.3"
pytz = "^2024.2"
python-dateutil = "^2.9.0.post0"
pydantic-settings = "^2.0.0" 

[tool.poetry.scripts]
start = "main:app" 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]  
testpaths = ["tests"]
python_files = ["test_*.py"]